server/pyproject.toml
[project]
name = "votelytics-api"
version = "0.1.0"
description = "Electoral predictions API for Votelytics"
authors = [
    {name = "Your Name"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "fastapi (>=0.120.2,<0.121.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pdfplumber (>=0.11.7,<0.12.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
